Split weapon tabs into two files, WeaponTabs and SkinList
